Refactor of the IQueue/Queue/QueueTest classes in order to make them more streamlined and more consistent with the rest of the QueueService. Tests also no longer rely on real consumers, so this should make them more stable.
